requirements.txt
appnope==0.1.0
decorator==4.0.6
Django==1.8.6
djangorestframework==3.3.1
gnureadline==6.3.3
ipdb==0.8.1
ipython==4.0.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
simplegeneric==0.8.1
traitlets==4.1.0
wheel==0.24.0
